Scroll during keyboard dragging
